parker_jpc2023/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==23.1.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.12.2
bleach==6.0.0
bunch==1.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastjsonschema==2.16.3
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.3
fqdn==1.5.1
idaes-pse==2.0.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.3
omlt==1.1
openpyxl==3.1.2
packaging==23.1
pandas==1.4.4
pandocfilters==1.5.0
-e git+https://github.com/idaes/publications.git@83bc504fbcf3a4b1fce882c482a3189a42a53842#egg=parker_jpc2023&subdirectory=parker_jpc2023
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
Pint==0.20.1
platformdirs==3.2.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.15.1
Pyomo==6.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
rbfopt==4.2.5
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sympy==1.11.1
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tinydb==4.7.1
tomli==2.0.1
tornado==6.3.1
traitlets==5.9.0
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
xlrd==2.0.1
zipp==3.15.0
